Reorder modifiers
Previously it was difficult to use repeated shortcuts like C-d and C-p
because the same finger had to be used for both. This meant it wasn't
possible to hold the control key and tap the letter, instead forcing
you to tap control then tap the letter.

Using this modifier order fixes the issue by moving control to the
pinky, allowing it to be held for C-d and C-p.